apps/farmbase/pyproject.toml
[project]
name = "farmbase"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = "<=3.13"
dependencies = [
    "aiohttp>=3.11.18",
    "alembic>=1.15.1",
    "arcgis>=2.4.1",
    "asyncpg>=0.30.0",
    "email-validator>=2.2.0",
    "emails>=0.6",
    "fastapi[standard]>=0.115.8",
    "geoalchemy2>=0.17.1",
    "geojson-pydantic>=1.2.0",
    "geopandas>=1.0.1",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.6.1",
    "pyjwt>=2.10.1",
    "python-jose>=3.4.0",
    "python-slugify>=8.0.4",
    "rioxarray>=0.19.0",
    "schemathesis>=3.39.16",
    "sentry-sdk>=2.23.1",
    "slowapi>=0.1.9",
    "sqlalchemy[asyncio]>=2.0.39",
    "sqlalchemy-filterset>=2.3.0",
    "sqlalchemy-utils>=0.41.2",
    "sqlmodel>=0.0.24",
    "tabulate>=0.9.0",
    "xvec>=0.3.1",
    "pydantic-extra-types>=2.10.4",
    "phonenumbers>=9.0.4",
    "loguru>=0.7.3",
    "temporalio>=1.11.1",
    "xarray[complete]>=2025.3.1",
    "python-weather>=2.0.7",
]

[dependency-groups]
dev = [
    "alembic-postgresql-enum>=1.7.0",
    "pdfplumber>=0.11.6",
    #    "basedpyright>0",
    "poethepoet>0",
    "pymupdf>=1.25.5",
    "pytest>0",
    "ruff>0",
]

[project.entry-points."farmbase.plugins"]
farmbase_basic_auth = "farmbase.plugins.farmbase_core.plugin:BasicAuthProviderPlugin"
farmbase_header_auth = "farmbase.plugins.farmbase_core.plugin:HeaderAuthProviderPlugin"

[tool.uv.sources]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
